+++ log10/llm.py
@@ -3,7 +3,6 @@ import logging
 import json
 from typing import List, Optional
 from abc import ABC, abstractmethod
-from copy import deepcopy
 from enum import Enum
 import os
 import sys
@@ -251,12 +250,3 @@ class NoopLLM(LLM):
     def text(self, prompt: str, hparams: dict = None) -> TextCompletion:
         logging.info("Received text completion requst: " + prompt)
         return TextCompletion(text="I'm not a real LLM")
-
-
-def merge_hparams(override, base):
-    merged = deepcopy(base)
-    if override:
-        for hparam in override:
-            merged[hparam] = override[hparam]
-
-    return merged
