+++ log10/together.py
@@ -4,6 +4,7 @@ import together
 
 from copy import deepcopy
 from log10.llm import LLM, Kind, Log10Config, TextCompletion
+from log10.utils import merge_hparams
 
 together.api_key = os.environ.get("TOGETHER_API_KEY")
 
@@ -41,7 +42,6 @@ class Together(LLM):
     def text(self, prompt: str, hparams: dict = None) -> TextCompletion:
         request = self.text_request(prompt, hparams)
         openai_request = {
-            "prompt": prompt,
             **request,
         }
 
@@ -62,10 +62,7 @@ class Together(LLM):
         return TextCompletion(text=response["choices"][0]["text"], response=response)
 
     def text_request(self, prompt: str, hparams: dict = None) -> dict:
-        merged_hparams = deepcopy(self.hparams)
-        if hparams:
-            for hparam in hparams:
-                merged_hparams[hparam] = hparams[hparam]
+        merged_hparams = merge_hparams(hparams, self.hparams)
         return {"prompt": prompt, **merged_hparams}
 
     def _prepare_response(self, completion: dict) -> dict:
